chore: Rename MultiProvider sample to MultiLock.MultiProvider
Refactor the multi-provider demo sample to align with the MultiLock naming convention:

- Rename project from LeaderElection.MultiProvider to MultiLock.MultiProvider
- Extract DemoBackgroundService into separate file
- Update project references to use MultiLock namespace (MultiLock.csproj, MultiLock.InMemory, MultiLock.FileSystem)
- Update using statements to match new namespace structure
- Update console output and temp path to use MultiLock branding
- Add project to solution file
- Apply modern C# syntax improvements (var to explicit types, simplified method calls)
